// src/typings/user.d.ts
export type MemberBaseSettingsInfoResp = {
  businessId: number // b 端 id
  uid: number // 会员 uid
  pushLanguage: string // 推送语言
  marketSetting: number // 涨跌色
  currencyTypeCd: string // 折算货币
  antiPhishingCode: string // 反钓鱼码
  isOpenActiveWarn: number // 是否启用最新活动提醒 1=true; 2=false
  isOpenContractStatus: number // 合约开通状态 1=true; 2=false
  isOpenCurrencyWarn: number // 是否启用新币提醒 1=true; 2=false
  isOpenOrderWarn: number // 是否启用合约订单提醒 1=true; 2=false
  isOpenPriceWarn: number // 是否启用行情异动提醒 1=true; 2=false
  isOpenSpotStatus: number // 杠杆现货开通状态 1=true; 2=false
  perpetualVersion: number // 合约版本，1，专业版，2，普通版
}
// ...
export type MemberUserInfoResp = UserInfoType

/** 刷新 token */
export type MemberAuthRefreshTokenReq = {
  refreshToken: string // 登陆成功后获取的 refreshToken 值
  tokenTtl: number
}

export type MemberAuthRefreshTokenResp = {
  accessToken: string // 新的 token
  refreshToken: string
  accessTokenExpireTime: number
  refreshTokenExpireTime: number
}